backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dlib>=19.24.9",
    "eisp>=0.3.5",
    "facenet-pytorch>=2.5.3",
    "fastapi[standard]>=0.136.0",
    "gdown>=5.2.0",
    "imutils>=0.5.4",
    "ipykernel>=6.30.0",
    "ipython>=8.37.0",
    "mtcnn>=1.0.0",
    "nbconvert>=7.16.6",
    "nbformat>=5.10.4",
    "numpy<2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.2.0",
    "polars>=1.34.0",
    "python-multipart>=0.0.26",
    "ruff>=0.11.2",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.0",
    "shap>=0.49.1",
    "tensorflow==2.11",
    "torch>=2.0.1",
    "torchvision>=0.15.2",
    "tqdm>=4.67.1",
    "transformers>=4.50.0",
    "ultralytics>=8.3.133",
    "umap-learn>=0.5.9.post2",
    "vl-convert-python>=1.8.0",
    "wget>=3.2",
    "xgboost==3.0.5",
]
